Updating BST determination code to use NSIntegerMin and NSIntegerMax

+++ Algorithms/Graphs/BinaryTree.h
@@ -13,6 +13,6 @@
 
 @interface BinaryTree : NSObject
 - (id)initWithRoot:(BinaryTreeNode *)root;
-- (BOOL)isBST:(BinaryTreeNode *)node withMin:(int)min andMax:(int)max;
+- (BOOL)isBST:(BinaryTreeNode *)node withMin:(NSInteger)min andMax:(NSInteger)max;
 - (BOOL)isBST;
 @end
